requirements.txt
flask==2.3.3
ldclient-py==6.9.1
pytest==7.4.0